% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_error_report.R
\name{create_error_report}
\alias{create_error_report}
\title{Create an error report HTML file}
\usage{
create_error_report(file, error_msg, dataset_title = "Unknown")
}
\arguments{
\item{file}{Output file path}

\item{error_msg}{Error message}

\item{dataset_title}{Dataset title}
}
\description{
Create an error report HTML file
}
\keyword{internal}
