% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fallback_template.R
\name{create_fallback_template}
\alias{create_fallback_template}
\title{Create a fallback template when package template is missing}
\usage{
create_fallback_template(template_path)
}
\arguments{
\item{template_path}{Path where to create the template}
}
\description{
Create a fallback template when package template is missing
}
\keyword{internal}
