% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_named_result.R
\name{create_named_result}
\alias{create_named_result}
\title{Create named result list from TidyTuesday data}
\usage{
create_named_result(tt_data, file_names)
}
\arguments{
\item{tt_data}{TidyTuesday data object}

\item{file_names}{Character vector of file names}
}
\value{
Named list of cleaned datasets
}
\description{
Create named result list from TidyTuesday data
}
\keyword{internal}
