% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_file_names.R
\name{extract_file_names}
\alias{extract_file_names}
\title{Extract file names from TidyTuesday data object}
\usage{
extract_file_names(tt_data, preserve_original_names = FALSE)
}
\arguments{
\item{tt_data}{TidyTuesday data object}

\item{preserve_original_names}{Logical, fallback to original names}
}
\value{
Character vector of file names
}
\description{
Extract file names from TidyTuesday data object
}
\keyword{internal}
