% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_plot.R
\name{inspect_plot}
\alias{inspect_plot}
\title{Generate an inspectdf Visualization from TidyTuesday Data}
\usage{
inspect_plot(ttd, plot = c("type", "na", "cor", "num", "cat"))
}
\arguments{
\item{ttd}{A list returned by \code{get_tt_data()}.}

\item{plot}{A string specifying which plot to produce:
\code{"type"}, \code{"na"}, \code{"cor"}, \code{"num"}, or \code{"cat"}.}
}
\value{
A \code{ggplot2} plot, or a list of plots (for categorical and numeric
checks).
}
\description{
Based on a \code{ttd} object returned by \code{get_tt_data()}, this function determines
how to apply \code{inspectdf} plotting functions depending on the number and types
of datasets returned.
}
\examples{
ttd <- get_tt_data("Netflix Titles")
inspect_plot(ttd, plot = "type")

ttd2 <- get_tt_data("Space Launches")
inspect_plot(ttd2, plot = "num")

}
