% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the shrapports application}
\usage{
launch_app()
}
\description{
This function starts the Shiny application for generating
downloadable reports from data visualizations.
}
\examples{
\dontrun{
launch_app()
}
}
