% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report.R
\name{report_server}
\alias{report_server}
\title{Report Module Server}
\usage{
report_server(
  id,
  dataset_name_reactive,
  data_reactive,
  x_var_reactive,
  y_var_reactive,
  color_var_reactive = NULL,
  table_data = NULL
)
}
\arguments{
\item{id}{The module ID}

\item{dataset_name_reactive}{Reactive expression returning the dataset name}

\item{data_reactive}{Reactive expression returning the current dataset}

\item{x_var_reactive}{Reactive expression returning the x variable name}

\item{y_var_reactive}{Reactive expression returning the y variable name}

\item{color_var_reactive}{Reactive expression returning the color variable name}

\item{table_data}{Reactive expression returning the summary table data}
}
\value{
A Shiny server module
}
\description{
Report Module Server
}
