% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R
\name{table_server}
\alias{table_server}
\title{Table Module Server}
\usage{
table_server(
  id,
  data_reactive,
  x_var_reactive,
  y_var_reactive,
  color_var_reactive = NULL
)
}
\arguments{
\item{id}{The module ID}

\item{data_reactive}{Reactive expression returning the dataset}

\item{x_var_reactive}{Reactive expression returning the x variable}

\item{y_var_reactive}{Reactive expression returning the y variable}

\item{color_var_reactive}{Reactive expression returning the color variable}
}
\value{
A Shiny server module
}
\description{
Table Module Server
}
