% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_var_input.R
\name{var_input_server}
\alias{var_input_server}
\title{Variable Input Module Server}
\usage{
var_input_server(id)
}
\arguments{
\item{id}{The module ID}
}
\value{
A list of reactive values
}
\description{
Variable Input Module Server
}
